refactor: migrate terminal plugin from im/ to cli/ module
Move TerminalAdapter and TerminalPlugin from src/im/ to src/cli/ to align
with design doc which specifies terminal channel lives in CLI module.

Changes:
- Move src/im/terminal.rs → src/cli/terminal.rs
- Move src/im/terminal_tests.rs → src/cli/terminal_tests.rs
- Add pub mod terminal + terminal_tests to cli/mod.rs
- Remove terminal modules from im/mod.rs
- Update all import paths (daemon, cli/chat, plugin_tests)

Source: user
Type: refactor
