+++ src/im_adapter/plugin_tests.rs
@@ -9,7 +9,7 @@
 
 #[cfg(test)]
 mod tests {
-    use crate::im::terminal::TerminalPlugin;
+    use crate::cli::terminal::TerminalPlugin;
     use crate::im_adapter::code_block::{parse_content_segments, ContentSegment};
     use crate::im_adapter::platforms::feishu::{FeishuAdapter, FeishuPlugin};
     use crate::im_adapter::plugin::{IMPlugin, RenderedOutput};
